Split parity gate suites by domain
